Move remark-linker to backend
